CDH-20117 Revert changes that introduced binary incompatiblity between 0.98 and 0.96
HBASE-10452 unnecessarily changed the public stable api signature of Get|Scan.setTimeStamp to throw an IOException.
We restored the old behavior that swallowed the impossible to trigger exception but keep the new warning message.

HBASE-10339 restored a method to 0.98 that was initially removed.  The restoration had the incorrect method signature.
We return a Map now instead of a NavigableMap to restore compiled code compatbility. (the modified code would work if recompiled).

Reason: Backwards Compatibility
Author: Jonathan Hsieh
Ref: CDH-20117