HBASE-13471 Fix a possible infinite loop in doMiniBatchMutation
Reason: Bug
Author: Rajesh Nishtala
Ref: CDH-31217

Summary: in doMiniBatchMutation it is possible to get into an infinite loop when a query has a row that is not in the region. If the batch had an invalid row, the row lock further down the function would fail to acquire because it was an invalid row. However we'd catch the exception and improperly treat it as if we had not acquired the lock and then try acquiring the lock again. Thus once we got into this state we'd be stuck in an infinite loop. Worse yet, this infiite loop would occur with the readLock held. So any other opertaions such as doClose() would be locked out and stuck. The patch is to check whether the row is valid and short circuit the failure when it doesn't work.

Test Plan: IntegrationTestReplication would consistently fail when trying to disable large tables before the fix. After the test the tests pass consistently.

Reviewers: eclark

Subscribers: asameet

Differential Revision: https://reviews.facebook.net/D37437

Signed-off-by: Elliott Clark <eclark@apache.org>