HBASE-16630 Fragmentation in long running Bucket Cache
Bug Fix : Added Memory block type also to the list of
evictions that need to happen when there is a needForExtra

Improvement : Inorder to fix the non availabilty of Buckets and force
the movement of buckets to transformed sizes, whenever we encounter a
situation where an allocation cant be made for a BucketSize, we will
forcefully free the entire buckets that have least occupancy ratio. This
is the same strategy used by MemCached when they encounter a similar
issue going by the name 'Slab Calcification'. Only improvement is that
we use a heuristic to evict from the buckets that are least occupied
and also avoid the BucketSizes where there is a single Bucket

Signed-off-by: Ramkrishna <ramkrishna.s.vasudevan@intel.com>

(cherry picked from commit ea5ac3856cde31b17615bbf38f54f2808a5647ff)

Change-Id: I23be8258ca8d13c730c2f7f97b762e165130284a
Author: dvdreddy
Reason: Bug
Ref: CDH-53420